requirements.txt
python-binance==1.0.19
typer==0.9.0
pydantic==2.5.0
loguru==0.7.2
python-dotenv==1.0.0